requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blis==0.7.11
bokeh==3.3.0
branca==0.7.0
bs4==0.0.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpathlib==0.15.1
cloudpickle==3.0.0
comm==0.1.4
confection==0.1.3
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dask==2023.10.1
dask_labextension==7.0.0
debugpy==1.8.0
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
descartes==1.1.0
dill==0.3.7
distributed==2023.10.1
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0-py3-none-any.whl#sha256=d37c3d60738be409a04f83a51421075b94fc02efad99bb8e62ec9018b1f99cb0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
executing==2.0.0
faiss-cpu==1.7.4
fastjsonschema==2.18.1
fasttext==0.9.2
filelock==3.12.4
fiona==1.9.5
folium==0.15.0
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
funcy==2.0
gensim==4.3.2
geopandas==0.14.1
huggingface-hub==0.17.3
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
iso639==0.1.4
isoduration==20.11.0
jedi==0.19.1
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.9.1
jupyter_server_proxy==4.1.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kaleido==0.2.1
kiwisolver==1.4.5
KrovetzStemmer==0.8
langcodes==3.3.0
langdetect==1.0.7
language-data==1.1
lingua==4.15.0
lingua-language-detector==1.3.3
locket==1.0.0
loguru==0.7.2
lxml==4.9.3
Mako==1.2.4
marisa-trie==0.7.8
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
missingno==0.5.2
mistune==3.0.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.24.4
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pathos==0.3.1
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.18.0
pluggy==1.3.0
polib==1.2.0
pox==0.3.3
ppft==1.7.6.7
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.11.1
pycountry==22.3.5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyExecJS==1.5.1
Pygments==2.16.1
pyLDAvis==3.4.1
pyparsing==3.1.1
pyproj==3.6.1
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.10.6
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
segtok==1.5.11
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.2
simpervisor==1.0.0
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.1
spacy-alignments==0.9.1
spacy-langdetect==0.1.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.2
SQLAlchemy==2.0.21
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tblib==3.0.0
tenacity==8.2.3
terminado==0.17.1
textblob==0.17.1
thinc==8.2.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.34.0
translate-api==4.9.5
typer==0.9.0
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.6
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.2
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
wordcloud==1.9.2
xyzservices==2023.10.1
yake==0.4.8
yarl==1.9.2
zict==3.0.0
zipp==3.17.0
